LayoutTests/storage/indexeddb/modern/leak-1-expected.txt

This tests that certain IDB object relationships don't cause leaks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
Initial upgrade needed: Old version - 0 New version - 1
Issuing a simple request to the object store
Observing GC on the transaction and request.
PASS txObserver.wasCollected is true
PASS reqObserver.wasCollected is true
PASS versionChangeObserver.wasCollected is true
PASS dbObserver.wasCollected is true
PASS openRequestObserver.wasCollected is true
PASS objectStoreObserver.wasCollected is true
PASS successfullyParsed is true

TEST COMPLETE
